A bit more refactoring and documenting
